rewrite audio handle for new results.json
there's now a new function check_DTMF() that takes a string with the
DTMF decoded reference text as argument. The function stops capturing,
decodes the captured audio and compares it with the reference string.
Then stores the result in results.json.
stop_audiocapture() still exists but only stores the wav file with
unknown result, analogue to take_screenshot().