make mouse_hide more reliable
by retrying till we have an absolute mouse pointer. I had to reverse
engineer qemu's vnc implementation to figure where the problem is. But
qemu has no mouse registered at the moment we connect to VNC and so
mouse moves are ignored - but they trigger a check if there is a mouse.
So we need to try one more time, till we know qemu has a mouse - fortunately
it tells us