better UEFI firmware handling: check existence, Fedora support
There's a couple of problems with how UEFI firmware is handled
at present. If you pass the firmware location as UEFI_BIOS, the
check that the file actually exists gets skipped, because of
how things are ordered. Also, it will try and find a UEFI
firmware for you on x86_64 if BIOS and UEFI_BIOS aren't set,
but it only knows one possible location, the one in the package
for openSUSE. The packages from https://kraxel.org, commonly
used on Fedora, place a UEFI firmware somewhere else; this adds
support for that location (and makes it easy to support other
known locations as we find them). Finally, I made it possible
to pass an absolute path rather than one relative to
/usr/share/qemu, if you like. This also puts almost all the
handling in one place, leaving only the actual addition of the
qemu param to happen later (after the @params array is defined).