Easier detectable errors for non-needle failures
Both wait_serial and die '' were pretty hard to detect on the webui
as they only reveal themselves to the knowing as red screens with a
hint in the log file. That's very hard for beginners to grasp

So change that: In case of wait_serial and die we record a 'fail' text result
together with a screenshot.

See https://progress.opensuse.org/issues/6526