Allow folder separated test distribution collections
Introduce the
If a main.pm can not be found in the CASEDIR itself, it reverts to a new
approach trying to load the main.pm from a subfolder specified by DISTRI
within CASEDIR. This allows e.g. *opensuse* and *sle* flavors of SUSE
products to be maintained within one branch of test repository.

Same for loading "needles/" directory.

Prerequisite for https://progress.opensuse.org/issues/10162

DISTRI was actually unused but still checked for existance since
9df5fb6a891fb515d56604f8664d827ed89004a0.