Use stack introspection for sub name in log calls
The 'name' field was used inconsistently in calls of log_call, fctinfo and
others. To correct and further simplify the call stack can be read to gather
the current sub name without the need to explicitly specify it.

All changed subfunctions are only used within os-autoinst. Only the log output
is affected besides the simplification of code.