Output git hash of executed tests
If the test directory (CASEDIR) is a git repository the current commit hash is
saved just before the tests are loaded and output into the logs. The git
commit hash can be used for reference in later analysis, e.g. in openQA to
display the file not in the "most recent" state but the actual state the test
source was in at the time of execution.

If CASEDIR is not a git repo "UNKNOWN" is stored instead. A "dirty" state of
the repository is *not* detected.

Example output

```
12:29:34.6697 git hash of test distribution: c9421dc270a53d5f408e22607e437be8ae385349
12:29:34.6803 usingenv DESKTOP=textmode
12:29:34.6804 usingenv TEXTMODE=1
```