Disable too strict stall detection again
7653967c704b0812ee5f08103a06601405a9da05 introduced a timeout limit with
intentional crash on hit. As it looks like we are hitting this limit more
often than expected, still. Many tests would continue fine even though a stall
is detection somewhere in between.

If it is desired to put the intentional crash back on stall it should be made
optional so that we can enable it in "development" runs but just record and
continue in "productive" runs.