testapi: Add check_act_and_assert helper method
'check_act_and_assert' can be used wherever check_screen is used to conduct an
optional action before an assert_screen, e.g. to check for an optional close
dialog. The approach to check for multiple tags in one assert_screen and check
if an optional needle has been matched to conduct an optional action can help
save time as the check_screen call does not run into a timeout if the
optional action is not necessary to be executed.