add an option to upload_logs to avoid dying on failure
in Fedora we use upload_logs a lot in post-fail hooks, and in
some cases the file may not exist. For instance, anaconda
creates lots of log files, some of which may not exist in some
circumstances. We would like to just be able to try and upload
them all but not have the post-fail hook die if any of them
does not exist, instead of having to test for each file's
existence before uploading it, or have a bunch of conditionals
that magically know which log files will exist in which state.