Move bmwqemu functions (ab-)used in openSUSE tests to test API
The functions 'hashed_string' and 'diag' are currently used in
os-autoinst-distri-opensuse but only the testapi functions should be used. If
we need them we may as well just include them in the test API and subsequently
clean up tests (and then bmwqemu). 'random_string' sounds similar to
'hashed_string' but is only used within bmwqemu so only the doc has been
adapted but the function not moved.