consoles: Die on timeout when searching for custom status line entry
d7b93259 wrongly assumed that "neither buffer_ready nor buffer_full" is an
expected step during search for another search term and replaced a confess
call by warn. The current commit again replaces this by confess and tries
to improve the message. This step is actually hit when a search hits a
timeout when looking for a status line entry, e.g. when waiting for
"X E D I T".