Fix the backend exiting
exit is calling all kind of exit handlers including the one of libusb, which
tries to close some monitoring that it started on library load (very strange
concept). As we fork off the data from isotovideo, we inherit all this but
can't (and shouldn't) erase it in the childs. So use _exit

On top of that perl 5.24 introduced AUTOLOADing DESTROY, which causes funny
deadlocks - so have an explit DESTROY, even when no longer DESTROYing due
to _exit (but safer is safer)