when match is equal quality, prefer workaround needle
resolves openQA issue #779. When two needles both match with
the same match quality, we currently just take whichever sorts
first by name alphabetically. We would like to predictably
prefer workaround needles to non-workaround in this case, so
we rename `cmp_by_error__` to `cmp_by_error_type_` and have it
also consider the needle type and prefer workaround needles, if
the match quality is equal.