Write the complete check result to the result file
In the old code, write a dummy needle to the result file in interactive
mode seems wrong, it should write the complete check result from
check_asserted_screen to the result file. Also moved $rsp->{saveresult}
part to the timeout block but just do not set overall state, $rsp->{timeout}
must be 1 when $rsp->{saversult} was set to 1.

https://progress.opensuse.org/issues/13456