Make backend<->isotovideo communication more strict
Whenever we ask the backend something, we add a random string and whenever
we look at the answer, we verify it's presence. This way we can't get
answers/queries in wrong order as it happend during unclean shutdown. The
logs would appear like this:

09:46:18.2756 26372 signalhandler got TERM - loop 1
09:46:18.2790 26372 isotovideo failed
09:46:18.2791 26372 send_json {"cmd":"is_shutdown","json_cmd_token":"gwkssdzz"}
09:46:19.0769 26372 sysread {"rsp":{"result":""},"json_cmd_token":"CUlerWmc"}
ERROR: the token does not match - questions and answers not in the right order at os-autoinst/myjsonrpc.pm line 75.
        myjsonrpc::read_json(GLOB(0x14faa98), "gwkssdzz") called at os-autoinst/backend/driver.pm line 165
        backend::driver::_send_json(backend::driver=HASH(0x38a9708), HASH(0x3b47e18)) called at os-autoinst/isotovideo line 414
09:46:19.0781 26372 killing backend process 26376
09:46:19.0826 26372 done with backend process