don't sha1sum qcow assets (poo#13526) (#574)
it takes a long time and blocks the worker process, preventing
it from sending status updates to the server and making it
vulnerable to the 'dead worker' check. Since we're *also*
checksumming in the worker itself, this really isn't needed.

Combined with several other changes I'm submitting for openQA,
this should address poo#13526.