testapi: type_string: optionally check for screen changes (#584)
this adds a new param for `type_string`, which basically waits
for the screen to change every few characters. You can configure
how many characters to type per `wait_screen_change` with the
`wait_screen_change` parameter. I wrote this to try and mitigate
an issue we are having where Fedora tests occasionally seem to
exhaust the VM's input buffer and thus miss characters (or repeat
a character a bunch of times, if the VM happens to get a key_down
event but not the key_up event), but it seems generally useful
and may as well go in testapi, I think.

Of course this has limitations: it's no use if the screen will
be changing whether keypresses are received or not (e.g. some
constantly-changing element is present on screen, or a flashing
cursor that's large enough to be considered a 'change' by
`wait_screen_change`), and it's also no use if there is no
screen change in response to the typing (e.g. typing into some
password prompts). It also may cause tests to take quite a long
time to fail if they wind up trying to type without the input
box or whatever highlighted, so the keystrokes don't change the
screen; there'll be several `wait_screen_change` timeouts before
the function eventually returns and a subsequent needle fail or
whatever notices that the test went wrong. This can be mitigated
with sensible needle matching before calling the function.