isotovideo: don't upload disks if test failed (POO #13722)
At present, the "# mark hard disks for upload if test finished"
block runs even if the test *failed*: it's only skipped if
something went really wrong with test execution somehow (which
causes $r to be set to 1). This is a problem because commonly,
if the test failed, the SUT will not have been shut down
cleanly, so we'll run into the "ERROR: Machine not shut down"
case, which causes us to exit 1 (which in turn, on current
openQA git, causes the test to wind up as 'incomplete' rather
than 'failed').

`autotest::runalltests` actually returns either 0 or 1 to signal
whether a fatal test failed or not, but nothing was using this
return value at present; `run_all` only cared if `runalltests`
*died*, not if it returned 0 or 1.

So to improve things, let's have `run_all` care both whether
`runalltests` died and what it returned, and pass both bits of
information along to `isotovideo`, which can then use them both
to decide whether to run the disk upload block or not.

If anyone has a case where they want a test to upload its disk
images regardless of whether the test failed, we could tweak
this a bit more - say, if `STORE_ON_FAIL` var is set, we can
run the upload block but skip the 'is VM shut down?' check if
`$passed` is 0, or something like that.

Add some tests that cover all this `run_all` stuff.