Wiggle the mouse in case of duplicated mouse_hides
this is basically a qemu workaround, but qemu VNC will ignore two
requests to move the mouse to the same absolute position. Human users
don't notice that as they *move* the mouse, they don't jump. So we
need to simulate mouse moves at least to some degree.

The duplicated mouse_hides being ignored are basically in situations
like:

mouse_hide;
restart_X; # will set the mouse cursor into the middle
mouse_hide;

(the mouse will be still at the center of the screen)