Started PoC on saving the state of the machine
Added a mock save_state method on qemu driver (also is present on
backend/baseclass.pm) which stops the vm, uses a migration and
saves a compressed image of the running machine. Test should
explicitly call this method for now. It should be introduced in
testbase.pm and tests having at least fatal flag should call this.

WIP
* Verify that the upload will work.
* Add command line used to spawn the VM.
* Add tests on both sides, openQA and os-autoinst.