Add disk save method and freeze vm method on test api
Add save_storage_drive method to test api and freeze_vm method aswell,
while the first is needed in case the test writer would explicitly
require the test to save the hard drives present in the system, the
latter is needed to pause/freeze the vm, so that the memory dumps or
disk dumps are not affected, in case something in between happens.

Rename save_state to save_memory_dump to improve code readability.

Improve the documentation for methods mentioned above and some minor
code changes aswell.

Currently depends on bsc#1008148