Add a flock together with the rsync in a shared needle directory for needle caching (#630)
* Add a flock together with the rsync in a shared needle directory

The needles are being copied form SOURCENEEDLES or PRODUCTDIR/needles
to a shared directory just one level above the current working
directory, allowing all of the workers share needles and using the
blocking flock to ensure that only one worker at a time is tyring to
rsync through isotovideo

Error code 999 will be triggered when the
lock can't be obtained.

* Use CACHEDIRECTORY variable to enable needle caching

The isotovideo script will now trigger needle caching only if
CACHEDIRECTORY variable is set in workers.ini, either globally or per
worker. It will use OPENQA_HOSTNAME and DISTRI variables to create the
directory tree; if it is not possible, the process will die.