Add debug output to needle load (#622)
* Add debug output to needle load

As we have an increasing number of incomplete jobs that stall during the
first iteration of assert_screen, I want to know if this is just some
kind of cache problem on the worker - or a performance problem of the NFS
the needles are coming from. Locally a needle load is less than 0.01 seconds,
but I have the suspicion we will see a factor 100 in production ;(

If we can prove this with this debug output (that we can limit to slow loads
once we know what slow is), we have to consider preload strategies

* Copy away the needle directory before the actual test starts

This tries to work around the problem we see with
https://progress.opensuse.org/issues/14252

NFS is just too unrealiable as far as timing is concerned, so we need to
create a copy from NFS into pool as long as we can wait

* Add a flock together with the rsync in a shared needle directory for needle caching (#630)

* Add a flock together with the rsync in a shared needle directory

The needles are being copied form SOURCENEEDLES or PRODUCTDIR/needles
to a shared directory just one level above the current working
directory, allowing all of the workers share needles and using the
blocking flock to ensure that only one worker at a time is tyring to
rsync through isotovideo

Error code 999 will be triggered when the
lock can't be obtained.

* Use CACHEDIRECTORY variable to enable needle caching

The isotovideo script will now trigger needle caching only if
CACHEDIRECTORY variable is set in workers.ini, either globally or per
worker. It will use OPENQA_HOSTNAME and DISTRI variables to create the
directory tree; if it is not possible, the process will die.