Do not take screenshot when VNC connection is dead
POO#14618

With qemu backend we use '-no-shutdown' Qemu option which prevents qemu
process to exit after it's guest OS quits. In svirt backend we don't
have such option so after OS is shutdown, the VNC connection is
terminated with it's domain - we should not take a screenshot in such a
case.