Add OpenQA::Benchmark::Stopwatch
Whith this, we can already start measuring times for the write_image
method.

Still some work needs to be done in base_class every time we are calling
write_img, as at one point it will just generate too much debug output.

Add check to enable chrono if C++11 is present (DEBUG_IO is disabled by
default).

The motivation for this utility class is to have a simple and smart
way to track different timers and calculate the deltas quickly.

At the begining [1], the class was doing a similar work, but only could
track named timers (instead of laps), and an overall timer had to be set
up to get the full time, and leading to a code that could be a bit
confusing [2]. After some research I stumbled upon Benchmark::Stopwatch
[3] which seemed like a better approach, and after some conversations
it was recommended to copy Benchmark::Stopwatch, because it is unmantained
since 2007, but it still builds fine.

The class is still mising an 'elapsed' method, and an automatic call to
stop() method when summary() is being called, but overall works fine.

Changed baseclass.pm to reflect the right changes according to the new
Benchmark::Stopwatch class.

Fix Image conversion benchmark and Stopwatch tests.
Add Stopwatch to Makefile.am.
Fix formatting for tinycv_impl.cc so that debug output for DEBUG_IO can
be easily parsed.

[1] https://git.io/v1MiP
[2] https://git.io/v1MiQ
[3] https://metacpan.org/changes/distribution/Benchmark-Stopwatch