Make the videoencoder to convert to png
Changed videoencoder to make it write the png when it has finished
encoding a new frame.

For now png and ppm will coexist (but the ppm will be a png internally
too). The only problem so far i've encountered is that the last.png is
not updated as often as we would like to.

http://deimos.suse.de/tests/1232#live has/had a live execution and it
shown the problem described above, delay in the WebUI... which took at
one point almost a minute.

Timing marks for this will be the next step.