Write last.png even if screen is not changing
Change the logic a bit, so that we check if we need to write a last.png
even if the frames repeat. The live view may start at any time, so we
need to be prepared a last.png at any time

Also support NOVIDEO by starting the video encoder with -n, which
creates an empty video (i.e. not using CPU for video)