Even safer shutdown to get published disks
While we try to stop the backend process, we will receive a SIGCHLD.
During handling this we should make sure not to remove sockets, we're
possibly writing to. This happens all the time, but for some reason
more likely on aarch64. This wouldn't be such a big problem if this
wouldn't stop (silently) to stop isotovideo from publishing disks.

See https://progress.opensuse.org/issues/14000