testapi: In documentation recommend `assert_script_run` by default
Focus on the more high-level methods which should be used normally for testing
unless different behaviour is required where one would fallback to the more
low-level methods.

Also delete incorrect statement about serial port. Within the opensuse test
distribution `script_run` and derivatives are used a lot with direct use of
the serial port which is not wrong, just needs to be handled with care.