Remove duplication in script_run and assert_script_run
`assert_script_run` and `script_run` currently duplicate the
work to echo the exit code of the command to the serial port
and act on it. Instead, let's have `assert_script_run` just
take the exit code from `script_run` and die unless it's 0.

Note that this leaves `script_sudo` and `assert_script_sudo`
alone. Ideally I'd like to harmonize those too, but that's more
complex and debatable. I did add a note that `assert_script_sudo`
is unlikely to actually work with the default `script_sudo`.

I cleaned up the docs a bit, and renamed `script_runs` args
to match `assert_script_run`'s. And I made the fail message for
the assert functions a bit more accurate (it can die because the
command timed out as well as because it failed).