Add optional 'no_wait' parameter for check/assert_screen
The default interval of checking screens every 1 second is an unsuitable
choices in multiple occasions. When screens change without external
interaction within a certain time, for example a boot menue, it is safer to
check more than once per second. In other occasions the screen repeatedly
alternates in an even interval and the screen check is conducted at the same
interval but shifted in time so that always the "wrong" screen is catched. In
this case the screen should be checked with (nearly) no waiting in between to
ensure the correct screen is catched.

In the cases of os-autoinst-distri-opensuse the openSUSE Krypton live CD is
built with SUSE kiwi v>=8 which yields a boot screen timeout of 1-2 seconds
Checking only once per second is just not often enough to be able to check the
screen and conduct actions immediately afterwards, e.g. press a key to abort
the timeout.

The second example is checking a text entry field in oomath where the needle
covers the blinking text cursor to be sure the focus is on that field. The
test fails here repeatedly but obviously not every time because the needle
checking interval matches the blinking cursor interval but is of by a fraction
of a second so that only the instances of the cursor not being shown is
matched.

Related progress issue: https://progress.opensuse.org/issues/16038