Only abort on system stalls detected when explicitly asked for
This changes the previous behaviour of always aborting the test in case of a
system stall to only abort when explicitly selected.

A system stall is bad but we can not fix it in the scope of os-autoinst
because it is an infrastructure issue. Also, not every check_screen or
assert_screen with a stall within is necessary for us to die on. Same as a
manual tester that goes for a cup of coffee. If he misses a dialog on the
screen with a timeout, this would be bad, but in most cases when he returns to
check the screen either the SUT is still busy doing something or already
waiting at the state where the user would need to give new inputs and the
tester actually checks the screen against his reference.

Explicit checking for stalls can be selected on the calls of
`check|assert_screen` as well as for the whole test execution with the
`vars.json` variable `ABORT_ON_STALL`.

Manually verified with the help of `t/99-full-stack.t` checking for correct
settings of the flag based on the testapi calls.