Reenable ZLRE encoding by default again
I disabled ZLRE because it's problematic both on iDRAC and on some
libvirt instances - and for qemu we don't really need the extra
compression.

But it turned out, both the zVM (SUT) and the zKVM (qemu) VNC
connections are problematic with the raw encoding - while it works
just fine with ZLRE. So reenable ZLRE by default - and disable it
explicitly for dell. On the libvirt instances, the problem seems
to be different anyway.

Details about the zVM can be found on
https://progress.opensuse.org/issues/16346