Only abort test on stall if assert_screen failed (not check)
As discussed in #713, this code which aborts the test if a stall
is detected during a needle check has a problem. It will abort
the test even in `check_screen`, not just `assert_screen` - so
it can cause tests to fail which otherwise would not have.

This makes `check_asserted_screen` only register whether a stall
was detected or not, and makes `_check_backend_response` fail
the test if a stall was detected *and* it's in 'assert mode'. If
it's in 'check mode', it just logs a warning about the stall,
but will continue the test.