Forward custom arguments to console_selected callback
Follow-up to https://github.com/os-autoinst/os-autoinst/pull/738

Needed for example when the callback is a check for the console to
be shown but the check should be done explicitly in the test. In this case
arguments can be passed through 'select_console' so that the distribution
specific callback can be called parameterized.

Related progress issue: https://progress.opensuse.org/issues/14826