Copy HDD & ISO images from NFS to host storage (#742)
poo#17824

Images are either created or copied to local storage. When they are
copied they are either used as they are (ISO images) or used as backing
file (e.g. JeOS & CaaSP).

`copy` parameter to add_disk() is now ignored and corretly channeled as
it used to. New `backingfile` parameter is used for backing files after
the file was copied.

Velidation runs (VM starts):
* Xen PV JeOS: http://assam.suse.cz/tests/5500
* Xen HVM JeOS: http://assam.suse.cz/tests/5496
* Xen PV SLES: http://assam.suse.cz/tests/5498
* Xen HVM SLES: http://assam.suse.cz/tests/5497