In case of stalled VNC relogin instead of showing black
The black screen was good enough for the use case we implemented it for (shut down
of s390 guest), but is very confusing in general. Now we try to relogin for 10 seconds
and will die if this fails. The backend is blocked during this period, so assert_screen
will not continue until we have a proper new picture

Tests better do not assert_screen on 'now rebooted' on platforms where the reboot drops
the VNC connection, but rather switch to a console that allows e.g. serial waits