videoencoder: Reduce number of allocations and do not copy data around
vector<uchar> is guranteed to be contigous memory in C++03 and later,
so it can be initialized directly from fread, saving the copy. Creating
the buf in the outer scope allows it to be allocated once and be reused.

Passing last_frame_image as optional imdecode argument allows the
underlying buffer to be reused.