tinycv: Blur only ROI instead of whole image
Most needles are significantly smaller than the whole screen, blur only
the region of interest.

The ROI is not used for the color conversion for two reasons - the blur
needs a slightly larger area for the convolution kernel, and the conversion
only contributes to the run time by a small fraction, thus creating a
fitting ROI is not worth the effort.

To avoid recreating the prepped image in case of multiple areas per needle,
merge the requested ROI with all earlier ROIs for the needle. In case the
new ROI is at least half of the images size, use the whole image, to
anticipate future ROIs.

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>