tinycv: restructure template search, use cv::minMaxLoc where possible
The search does not use the best match, but prefers "sufficient" matches
near the original location.

minMaxLoc only provides the best match, but provides information for
some shortcuts:
1. If the best match is bad, there is no reason to analyze any other
2. If the best match is at the original location, there is no reason to
  look for matches farther away
3. The best match provides an error upper bound for any reasonable match,
  and a maximum search radius
1. and 2. are actually the common cases for the needle search (total
mismatch due to wrong needle, or perfect match), and minMaxLoc is much
faster than the exhaustive search, it is reasonable to do a search for
the best match first.

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>