Extend run_cmd() to run on arbitrary server
Previously execution of commands via SSH was possible only on
`VIRSH_HOSTNAME`, whic is a virtualization server, a place where VM is
actually run. With Hyper-V, a platform without in-VM VNC support, we
use an intermediary Linux server, where xfreerdp can be funneled to VNC.
To execute command on intermedary which is not `VIRSH_HOSTNAME`
`run_cmd()` has to be extended by 'hostname' and 'password'.