serial_terminal: Wait for prompt (#797)
It appears that os-autoinst is sometimes too fast for the SUT and starts
sending characters before the shell (e.g. Bash) is ready to receive
them. Possibly the input is sent to the previous command's STDIN before it has
chance to close the file descriptors or at least before Bash resumes
interpreting TTY input.

I have only seen this happen on the virtio_console, but it may also happen on
VNC if the SUT stalls at just the right time.

The solution is to wait for the terminal prompt to be printed. Possibly there
is still a race condition where the shell doesn't start accepting input for a
period after printing the prompt. However I have not seen any evidence of this
yet.