Make sure last received frame is passed on after some delay (#801)
The needle matching might work on outdated frames if the last changes
where below the threshold. Implement a leaky bucket scheme for the
threshold so all frames but exact repetitions are candidates for updates.

For the video, the error might even have had accumulated, e.g. by a series
of frames with a similarity more than 50.

Fix for poo#19302

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>