Add proxy and dns component with local unit tests
Two new components where introduced for handling and redirecting
http/dns requests. The two components works together but can be used
separately, their behavior can be tweaked from more-than-one variable.

The most important variables are:
- CONNECTIONS_HIJACK (required): when setted to 1, both components (DNS and Proxy)
are loaded into the backend.
- CONNECTIONS_HIJACK_DNS (optional): when setted to 1 enable the DNS component
- CONNECTIONS_HIJACK_PROXY (optional): when setted to 1 enable the Proxy component
- CONNECTIONS_HIJACK_DNS_ENTRY (optional): can be used to supply a comma separated
list of host:ip domain to resolve to. It supports fine-grained policy
per-domain (instead of ip, you can drop all DNS requests to the domain
by setting 'DROP', or as for the opposite, with 'FORWARD' will be
forwarded to an external DNS server)
- CONNECTIONS_HIJACK_PROXY_ENTRY (optional): can be used to supply a comma
separated list of hosts which should be redirected to another one, in
form of host:redirhost. It can be supplied also a list of rewrite url
rules (as regexp) that could be used to transform the url to be
redirected to.

It is also possible to supply PORT/ADDRESS of the services, but are
setted up automatically based on the VNC port (that is assigned from
openQA).

Related Progress issue: https://progress.opensuse.org/issues/17436