Wire up components loading in baseclass and qemu/virt backend
With this change we allow to dynamically load components. Components to
be loaded can be supplied in the baseclass/backend constructor with the
'components' argument.

Any backend can request components to be loaded in such way:

	$bmwqemu::backend = backend::driver->new(
        	backend    => $bmwqemu::vars{BACKEND},
        	components => [qw( foo bar )]); # Load foo and bar

Components are automatically detected if are in the backend::component::
namespace.

To accomplish this, it was needed to refactor contructor code for
baseclass, qemu and svirt ( will be required for each backend that is
overriding the default constructor ), thus allowing the dynamic loading of
components.