Unconditionally set serial id on qemu drives (bsc#1041692)
Normally one would expect a serial ID on each HDD drive. Previously we set
this only for NVME devices and not at all for other devices. To fix
bsc#1041692 we need to explicitly set it and it is making the case easier if
we just apply this in all cases. Probably we will not have a problem with that
in all currently used qemu versions.

Verification run: https://openqa.suse.de/tests/1044039