Add process package
The class is a first effort to refactorize duplicated code to handle
external processes.
It is exposing basic IPC functionalities in a uniform way: either we
want to run code in a forked process or an external program,
it allows to communicate and manage it in OOP paradigm. The class is
covered by unit tests, incrementing coverage on process management in
the codebase.
Hopefully, if we switch to this implementation, it should fix or at least
make easier to debug issues like https://progress.opensuse.org/issues/6958