Introduce check_recorded_sound method (#873)
We have unstable aplay test, which fails due to the bsc#1048271 once in
a while. We cannot work it around with needles, as pattern is different
event if match level is 60%. In order to soft-fail this test module, we
introduce check_recorded_sound method, which doesn't fail the test, but
returns match result.