Xen: Support multiple HDDs
To support RAID test on svirt, multiple-disk support needs to be added
to bootloaded_svirt. For this to work, we can't use IDE device, because
it seems that Xen devices are in presence of IDE device connected to IDE
controller - but we can't connect more than four devices to IDE
controller. So switching to CDROM to SCSI makes the rest of the disks to
attach to SCSI controller (SCSI limit is 64 devices).

Validation runs on Xen HVM:
* JeOS: http://assam.suse.cz/tests/1003
* CaaSP: http://assam.suse.cz/tests/1005
* SLE15 lvm+RAID1: http://assam.suse.cz/tests/993